move modules/references to modules/experiences
